requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
pydantic==2.5.0
pyyaml==6.0.1
aiofiles==23.2.1
httpx==0.25.2
aiohttp==3.9.1
click==8.1.7
pytest==7.4.3
pytest-asyncio==0.21.1
asyncio-mqtt==0.16.1
watchdog==3.0.0
